Updates based on field testing.
moved all sched3 items to sched 2. Eventually each instrument should get
a 1,2 or 3 in config.txt so we know where to log it.
Setting a schedule rate to 0 in config.txt disables it.
All DEBUG log messages will now be logged to a file.
The hydrospere will assume its on battery power only if it ever sees a
voltage over 5.0 volts. If not on battery power, it must be on USB and
won't try to trigger solenoid until it sees a battery. It wouldn't work
even if it tried due to low voltage. This will help tell the difference
between USB power and a dead battery.